backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
alembic==1.13.1
marshmallow==3.20.2
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7
requests==2.31.0
geopy==2.4.1
pandas==2.1.4
numpy==1.26.2
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
python-magic==0.4.27
matplotlib==3.8.2
seaborn==0.13.0